+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Fixed compatibility with Controlify 2.0 beta 9.
+- Added an automatic righting option in the config.